pyproject.toml
[tool.poetry]
name = "jolteon"
version = "0.1.0"
description = ""
authors = ["Alex Ceccotti <alexceccotti5995@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
jolteon = "jolteon.__main__:app"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.5.3"
psycopg2-binary = "^2.9.5"
pydantic = {extras = ["dotenv"], version = "^1.10.5"}
pyyaml = "^6.0"
tabulate = "^0.9.0"
typer = "^0.7.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = {extras = ["colors"], version = "^5.12.0"}
mypy = "^1.0.1"
pytest = "^7.2.1"
types-psycopg2 = "^2.9.21.6"
types-pyyaml = "^6.0.12.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "pandas.*",
]
ignore_missing_imports = true
